Claims repo should be internal, remove token auth

+++ tests/integration/user_management/claims_repository/test_api.py
@@ -65,15 +65,13 @@ def test_get_from_a_random_path(client):
     assert response.status_code == status.HTTP_404_NOT_FOUND
 
 
-def test_post_claim(client_with_db, steward_headers):
+def test_post_claim(client_with_db):
     """Test that creating a user claim works."""
 
     user_dao = DummyUserDao()
     client_with_db.app.dependency_overrides[get_user_dao] = lambda: user_dao
 
-    response = client_with_db.post(
-        "/users/john@ghga.de/claims", json=ROLE_CLAIM_DATA, headers=steward_headers
-    )
+    response = client_with_db.post("/users/john@ghga.de/claims", json=ROLE_CLAIM_DATA)
 
     claim = response.json()
     assert response.status_code == status.HTTP_201_CREATED, claim
@@ -91,38 +89,12 @@ def test_post_claim(client_with_db, steward_headers):
     assert claim == ROLE_CLAIM_DATA
 
     # test with non-existing user
-    response = client_with_db.post(
-        "/users/john@haag.de/claims", json=ROLE_CLAIM_DATA, headers=steward_headers
-    )
+    response = client_with_db.post("/users/john@haag.de/claims", json=ROLE_CLAIM_DATA)
     assert response.status_code == status.HTTP_404_NOT_FOUND
     assert response.json()["detail"] == "The user was not found."
 
 
-def test_post_claim_for_same_user(client, steward_headers):
-    """Test that creating a user claim for the same user does not work."""
-
-    response = client.post(
-        "/users/steve-internal/claims", json=ROLE_CLAIM_DATA, headers=steward_headers
-    )
-    assert response.status_code == status.HTTP_403_FORBIDDEN
-    assert response.json()["detail"] == "Cannot modify own claims."
-
-
-def test_post_claim_without_permission(client, no_steward_headers):
-    """Test that creating a user claim without permission does not work."""
-
-    response = client.post("/users/john@ghga.de/claims", json=ROLE_CLAIM_DATA)
-    assert response.status_code == status.HTTP_403_FORBIDDEN
-    assert response.json()["detail"] == "Not authenticated"
-
-    response = client.post(
-        "/users/john@ghga.de/claims", json=ROLE_CLAIM_DATA, headers=no_steward_headers
-    )
-    assert response.status_code == status.HTTP_403_FORBIDDEN
-    assert response.json()["detail"] == "Not authorized"
-
-
-def test_get_claims(client_with_db, steward_headers):
+def test_get_claims(client_with_db):
     """Test that getting all claims of a user works."""
 
     user_dao = DummyUserDao()
@@ -140,13 +112,11 @@ def test_get_claims(client_with_db, steward_headers):
     }
     posted_claims = []
     for claim in (claim1, claim2):
-        response = client_with_db.post(
-            f"/users/{user_id}/claims", json=claim, headers=steward_headers
-        )
+        response = client_with_db.post(f"/users/{user_id}/claims", json=claim)
         assert response.status_code == status.HTTP_201_CREATED
         posted_claims.append(response.json())
 
-    response = client_with_db.get(f"/users/{user_id}/claims", headers=steward_headers)
+    response = client_with_db.get(f"/users/{user_id}/claims")
     assert response.status_code == status.HTTP_200_OK
     requested_claims = response.json()
     requested_claims.sort(key=itemgetter("assertion_date"))
@@ -160,37 +130,12 @@ def test_get_claims(client_with_db, steward_headers):
     assert requested_claims[0]["visa_value"] != requested_claims[1]["visa_value"]
 
     # test with non-existing user
-    response = client_with_db.get("/users/john@haag.de/claims", headers=steward_headers)
+    response = client_with_db.get("/users/john@haag.de/claims")
     assert response.status_code == status.HTTP_404_NOT_FOUND
     assert response.json()["detail"] == "The user was not found."
 
 
-def test_get_claims_for_same_user(client_with_db, steward_headers):
-    """Test that getting user claims for the same user is not forbidden."""
-
-    user_dao = DummyUserDao(id_="steve-internal")
-    client_with_db.app.dependency_overrides[get_user_dao] = lambda: user_dao
-
-    response = client_with_db.get(
-        "/users/steve-internal/claims", headers=steward_headers
-    )
-    assert response.status_code == status.HTTP_200_OK
-    assert response.json() == []
-
-
-def test_get_claims_without_permission(client, no_steward_headers):
-    """Test that getting user claims without permission does not work."""
-
-    response = client.get("/users/john@ghga.de/claims")
-    assert response.status_code == status.HTTP_403_FORBIDDEN
-    assert response.json()["detail"] == "Not authenticated"
-
-    response = client.get("/users/john@ghga.de/claims", headers=no_steward_headers)
-    assert response.status_code == status.HTTP_403_FORBIDDEN
-    assert response.json()["detail"] == "Not authorized"
-
-
-def test_patch_claim(client_with_db, steward_headers):
+def test_patch_claim(client_with_db):
     """Test that revoking a user claim works."""
 
     user_dao = DummyUserDao()
@@ -199,9 +144,7 @@ def test_patch_claim(client_with_db, steward_headers):
     user_id = "john@ghga.de"
 
     # post test claim
-    response = client_with_db.post(
-        f"/users/{user_id}/claims", json=ROLE_CLAIM_DATA, headers=steward_headers
-    )
+    response = client_with_db.post(f"/users/{user_id}/claims", json=ROLE_CLAIM_DATA)
     posted_claim = response.json()
     assert response.status_code == status.HTTP_201_CREATED
 
@@ -213,12 +156,12 @@ def test_patch_claim(client_with_db, steward_headers):
     revocation_date = "2022-10-15T12:00:00+00:00"
     patch_data = {"revocation_date": revocation_date}
     response = client_with_db.patch(
-        f"/users/{user_id}/claims/{claim_id}", json=patch_data, headers=steward_headers
+        f"/users/{user_id}/claims/{claim_id}", json=patch_data
     )
     assert response.status_code == status.HTTP_204_NO_CONTENT
 
     # test that claim has been revoked
-    response = client_with_db.get(f"/users/{user_id}/claims", headers=steward_headers)
+    response = client_with_db.get(f"/users/{user_id}/claims")
     assert response.status_code == status.HTTP_200_OK
     claims = response.json()
     assert len(claims) == 1
@@ -229,14 +172,14 @@ def test_patch_claim(client_with_db, steward_headers):
     # test without revocation date
     patch_data = {"revocation_date": None}  # type: ignore
     response = client_with_db.patch(
-        f"/users/{user_id}/claims/{claim_id}", json=patch_data, headers=steward_headers
+        f"/users/{user_id}/claims/{claim_id}", json=patch_data
     )
     assert response.status_code == status.HTTP_422_UNPROCESSABLE_ENTITY
 
     # test with later revocation date
     patch_data = {"revocation_date": "2022-10-15T13:00:00+00:00"}
     response = client_with_db.patch(
-        f"/users/{user_id}/claims/{claim_id}", json=patch_data, headers=steward_headers
+        f"/users/{user_id}/claims/{claim_id}", json=patch_data
     )
     assert response.status_code == status.HTTP_422_UNPROCESSABLE_ENTITY
     assert response.json()["detail"] == "Already revoked earlier."
@@ -245,12 +188,12 @@ def test_patch_claim(client_with_db, steward_headers):
     revocation_date = "2022-10-15T11:00:00+00:00"
     patch_data = {"revocation_date": revocation_date}
     response = client_with_db.patch(
-        f"/users/{user_id}/claims/{claim_id}", json=patch_data, headers=steward_headers
+        f"/users/{user_id}/claims/{claim_id}", json=patch_data
     )
     assert response.status_code == status.HTTP_204_NO_CONTENT
 
     # test that revocation was advanced
-    response = client_with_db.get(f"/users/{user_id}/claims", headers=steward_headers)
+    response = client_with_db.get(f"/users/{user_id}/claims")
     assert response.status_code == status.HTTP_200_OK
     claims = response.json()
     assert len(claims) == 1
@@ -260,55 +203,19 @@ def test_patch_claim(client_with_db, steward_headers):
 
     # test with non-existing user
     response = client_with_db.patch(
-        f"/users/john@haag.de/claims/{claim_id}",
-        json=patch_data,
-        headers=steward_headers,
+        f"/users/john@haag.de/claims/{claim_id}", json=patch_data
     )
     assert response.status_code == status.HTTP_404_NOT_FOUND
     assert response.json()["detail"] == "The user was not found."
     # test with non-existing claim
     response = client_with_db.patch(
-        f"/users/{user_id}/claims/invalid-claim-id",
-        json=patch_data,
-        headers=steward_headers,
+        f"/users/{user_id}/claims/invalid-claim-id", json=patch_data
     )
     assert response.status_code == status.HTTP_404_NOT_FOUND
     assert response.json()["detail"] == "The user claim was not found."
 
 
-def test_patch_claim_for_same_user(client, steward_headers):
-    """Test that modifying a user claim for the same user does not work."""
-
-    revocation_date = "2000-01-01T00:00:00+00:00"
-    patch_data = {"revocation_date": revocation_date}
-    response = client.patch(
-        "/users/steve-internal/claims/some-claim",
-        json=patch_data,
-        headers=steward_headers,
-    )
-    assert response.status_code == status.HTTP_403_FORBIDDEN
-    assert response.json()["detail"] == "Cannot modify own claims."
-
-
-def test_patch_claim_without_permission(client, no_steward_headers):
-    """Test that modifying a user claim without permission does not work."""
-
-    revocation_date = "2000-01-01T00:00:00+00:00"
-    patch_data = {"revocation_date": revocation_date}
-    response = client.patch("/users/some-user/claims/some-claim", json=patch_data)
-    assert response.status_code == status.HTTP_403_FORBIDDEN
-    assert response.json()["detail"] == "Not authenticated"
-
-    response = client.patch(
-        "/users/some-user/claims/some-claim",
-        json=patch_data,
-        headers=no_steward_headers,
-    )
-    assert response.status_code == status.HTTP_403_FORBIDDEN
-    assert response.json()["detail"] == "Not authorized"
-
-
-def test_delete_claim(client_with_db, steward_headers):
+def test_delete_claim(client_with_db):
     """Test that deleting a user claim works."""
 
     user_dao = DummyUserDao()
@@ -321,9 +228,7 @@ def test_delete_claim(client_with_db, steward_headers):
     claim2 = ROLE_CLAIM_DATA.copy()
     claim2["visa_type"] = "ResearcherStatus"
     for claim in (claim1, claim2):
-        response = client_with_db.post(
-            f"/users/{user_id}/claims", json=claim, headers=steward_headers
-        )
+        response = client_with_db.post(f"/users/{user_id}/claims", json=claim)
         assert response.status_code == status.HTTP_201_CREATED
         claim_id = response.json()["id"]
         assert claim_id
@@ -331,25 +236,21 @@ def test_delete_claim(client_with_db, steward_headers):
 
     # test deletion of first claim with non-existing user
     claim_id = claim1["id"]
-    response = client_with_db.delete(
-        f"/users/john@haag.de/claims/{claim_id}", headers=steward_headers
-    )
+    response = client_with_db.delete(f"/users/john@haag.de/claims/{claim_id}")
     assert response.status_code == status.HTTP_404_NOT_FOUND
 
     # test that claims have been posted
-    response = client_with_db.get(f"/users/{user_id}/claims", headers=steward_headers)
+    response = client_with_db.get(f"/users/{user_id}/claims")
     assert response.status_code == status.HTTP_200_OK
     claims = response.json()
     assert len(claims) == 2
 
     # delete first claim properly
-    response = client_with_db.delete(
-        f"/users/{user_id}/claims/{claim_id}", headers=steward_headers
-    )
+    response = client_with_db.delete(f"/users/{user_id}/claims/{claim_id}")
     assert response.status_code == status.HTTP_204_NO_CONTENT
 
     # test that claim has been deleted
-    response = client_with_db.get(f"/users/{user_id}/claims", headers=steward_headers)
+    response = client_with_db.get(f"/users/{user_id}/claims")
     assert response.status_code == status.HTTP_200_OK
     claims = response.json()
     assert len(claims) == 1
@@ -358,56 +259,26 @@ def test_delete_claim(client_with_db, steward_headers):
     assert claim["visa_type"] == "ResearcherStatus"
 
     # delete again
-    response = client_with_db.delete(
-        f"/users/{user_id}/claims/{claim_id}", headers=steward_headers
-    )
+    response = client_with_db.delete(f"/users/{user_id}/claims/{claim_id}")
     assert response.status_code == status.HTTP_404_NOT_FOUND
 
     # delete second claim
     claim_id = claim2["id"]
-    response = client_with_db.delete(
-        f"/users/{user_id}/claims/{claim_id}", headers=steward_headers
-    )
+    response = client_with_db.delete(f"/users/{user_id}/claims/{claim_id}")
     assert response.status_code == status.HTTP_204_NO_CONTENT
 
     # test that claim has been deleted
-    response = client_with_db.get(f"/users/{user_id}/claims", headers=steward_headers)
+    response = client_with_db.get(f"/users/{user_id}/claims")
     assert response.status_code == status.HTTP_200_OK
     claims = response.json()
     assert len(claims) == 0
 
     # delete again
-    response = client_with_db.delete(
-        f"/users/{user_id}/claims/{claim_id}", headers=steward_headers
-    )
+    response = client_with_db.delete(f"/users/{user_id}/claims/{claim_id}")
     assert response.status_code == status.HTTP_404_NOT_FOUND
 
 
-def test_delete_claim_for_same_user(client, steward_headers):
-    """Test that deleting a user claim for the same user does not work."""
-
-    response = client.delete(
-        "/users/steve-internal/claims/some-claim", headers=steward_headers
-    )
-    assert response.status_code == status.HTTP_403_FORBIDDEN
-    assert response.json()["detail"] == "Cannot modify own claims."
-
-
-def test_delete_claim_without_permission(client, no_steward_headers):
-    """Test that deleting a user claim without permission does not work."""
-
-    response = client.delete("/users/some-user/claims/some-claim")
-    assert response.status_code == status.HTTP_403_FORBIDDEN
-    assert response.json()["detail"] == "Not authenticated"
-
-    response = client.delete(
-        "/users/some-user/claims/some-claim", headers=no_steward_headers
-    )
-    assert response.status_code == status.HTTP_403_FORBIDDEN
-    assert response.json()["detail"] == "Not authorized"
-
-
-def test_check_download_access(client_with_db, steward_headers):
+def test_check_download_access(client_with_db):
     """Test that checking download access for a single dataset works."""
 
     user_dao = DummyUserDao()
@@ -420,9 +291,7 @@ def test_check_download_access(client_with_db, steward_headers):
     claim_data["valid_from"] = current_timestamp
     claim_data["valid_until"] = current_timestamp + 60
 
-    response = client_with_db.post(
-        "/users/john@ghga.de/claims", json=claim_data, headers=steward_headers
-    )
+    response = client_with_db.post("/users/john@ghga.de/claims", json=claim_data)
 
     claim = response.json()
     assert response.status_code == status.HTTP_201_CREATED
@@ -437,9 +306,7 @@ def test_check_download_access(client_with_db, steward_headers):
     claim_data["valid_from"] = current_timestamp - 60
     claim_data["valid_until"] = current_timestamp - 30
 
-    response = client_with_db.post(
-        "/users/john@ghga.de/claims", json=claim_data, headers=steward_headers
-    )
+    response = client_with_db.post("/users/john@ghga.de/claims", json=claim_data)
 
     claim = response.json()
     assert response.status_code == status.HTTP_201_CREATED
@@ -481,7 +348,7 @@ def test_check_download_access(client_with_db, steward_headers):
     assert response.json() == "false"
 
 
-def test_get_datasets_with_download_access(client_with_db, steward_headers):
+def test_get_datasets_with_download_access(client_with_db):
     """Test that getting all datasets with download access works."""
 
     user_dao = DummyUserDao()
@@ -500,9 +367,7 @@ def test_get_datasets_with_download_access(client_with_db, steward_headers):
     claim_data["valid_from"] = current_timestamp
     claim_data["valid_until"] = current_timestamp + 60
 
-    response = client_with_db.post(
-        "/users/john@ghga.de/claims", json=claim_data, headers=steward_headers
-    )
+    response = client_with_db.post("/users/john@ghga.de/claims", json=claim_data)
 
     claim = response.json()
     assert response.status_code == status.HTTP_201_CREATED
@@ -515,9 +380,7 @@ def test_get_datasets_with_download_access(client_with_db, steward_headers):
 
     claim_data["visa_value"] = claim_data["visa_value"].replace("some", "another")
 
-    response = client_with_db.post(
-        "/users/john@ghga.de/claims", json=claim_data, headers=steward_headers
-    )
+    response = client_with_db.post("/users/john@ghga.de/claims", json=claim_data)
 
     claim = response.json()
     assert response.status_code == status.HTTP_201_CREATED
@@ -532,9 +395,7 @@ def test_get_datasets_with_download_access(client_with_db, steward_headers):
     claim_data["valid_from"] = current_timestamp - 60
     claim_data["valid_until"] = current_timestamp - 30
 
-    response = client_with_db.post(
-        "/users/john@ghga.de/claims", json=claim_data, headers=steward_headers
-    )
+    response = client_with_db.post("/users/john@ghga.de/claims", json=claim_data)
 
     claim = response.json()
     assert response.status_code == status.HTTP_201_CREATED
@@ -559,12 +420,10 @@ def test_get_datasets_with_download_access(client_with_db, steward_headers):
     assert sorted(dataset_ids) == ["another-dataset-id", "some-dataset-id"]
 
 
-def test_get_claims_for_seeded_data_steward(client_with_db, steward_headers):
+def test_get_claims_for_seeded_data_steward(client_with_db):
     """Test that the database is seeded with the configured data steward."""
 
-    response = client_with_db.get(
-        "/users/the-id-of-rod-steward/claims", headers=steward_headers
-    )
+    response = client_with_db.get("/users/the-id-of-rod-steward/claims")
     assert response.status_code == status.HTTP_200_OK
     claims = response.json()
     assert len(claims) == 1
